HaikuDepot: Icon / Tarball Handling
Instead of exploding the tar-ball of icons from
HDS, keep it as a tar-ball, index it and read
data directly out from specific indicies on demand.

This will speed up the process of downloading
the icons by removing the unpack.  Also updates
will be faster by avoiding the need to delete the
old icon files.

Because icons are loaded on-demand, the start
time is faster by avoiding all the icon loads.
There are also savings on memory consumption.

Indexing on each load is surprisingly fast so
no external index is maintained.  Likewise for
the tar-balls's meta-data.

This commit does not cover the implementation
of an LRU cache of the icons in memory.

Relates to #15370

+++ src/apps/haikudepot/server/AbstractSingleFileServerProcess.cpp
@@ -42,7 +42,7 @@ AbstractSingleFileServerProcess::RunInternal()
 		result = DeleteLocalFile(localPath);
 
 	bool hasData = false;
-	off_t size;
+	size_t size;
 
 	if (IsSuccess(result))
 		result = StorageUtils::ExistsObject(localPath, &hasData, NULL, &size);
